+++ internal/debug/round_tripper_release.go
@@ -8,5 +8,9 @@ import "net/http"
 // RoundTripper returns a new http.RoundTripper which logs all requests (if
 // debug is enabled). When debug is not enabled, upstream is returned.
 func RoundTripper(upstream http.RoundTripper) http.RoundTripper {
+	if opts.isEnabled {
+		// only use loggingRoundTripper if the debug log is configured
+		return loggingRoundTripper{eofDetectRoundTripper{upstream}}
+	}
 	return upstream
 }
