+++ internal/debug/round_tripper_test.go
@@ -1,5 +1,3 @@
-// +build debug
-
 package debug
 
 import (
